refactor(app-headless-cms): rename Styled to StyledComponents
