+++ packages/app-headless-cms/src/admin/plugins/permissionRenderer/components/StyledComponents.tsx
